+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.0
+* [enhance] - enhance StickyCollapsablePanelController for command style control of collapse/expand the panel
+* [break change] - move parameter `disableCollapsable` and `defaultExpanded` to panelController
+* [break change] - rename parameter `controller` to `panelController` for better readability
+
 ## 3.0.4
 * [fix] - fix header expand/collapse state not sync when header size change #3